refactored tm handling
