+++ src/tmtccmd/core/frontend.py
@@ -17,7 +17,7 @@ from PyQt5.QtGui import QPixmap, QIcon
 from tmtccmd.core.backend import TmTcHandler
 from tmtccmd.core.definitions import CoreComInterfaces, CoreGlobalIds, CoreModeList
 from tmtccmd.config.globals import get_global_apid
-from tmtccmd.pus_tc.base import PusTelecommand
+from tmtccmd.pus_tc.definitions import PusTelecommand
 from tmtccmd.utility.tmtcc_logger import get_logger
 from tmtccmd.core.globals_manager import get_global, update_global
 from tmtccmd.com_if.tcpip_utilities import TcpIpConfigIds
