+++ src/tmtccmd/defaults/tc_packing.py
@@ -1,7 +1,7 @@
 import os
 from collections import deque
 
-from tmtccmd.pus_tc.base import TcQueueT
+from tmtccmd.pus_tc.definitions import TcQueueT
 from tmtccmd.core.definitions import CoreServiceList
 from tmtccmd.pus_tc.service_17_test import pack_service17_ping_command
 from tmtccmd.pus_tc.service_5_event import pack_generic_service5_test_into
