+++ src/tmtccmd/pus_tc/packer.py
@@ -8,7 +8,7 @@ This file transfers TC packing to the user application.
 """
 import sys
 
-from tmtccmd.pus_tc.base import TcQueueT
+from tmtccmd.pus_tc.definitions import TcQueueT
 from tmtccmd.utility.tmtcc_logger import get_logger
 
 LOGGER = get_logger()
