+++ src/tmtccmd/pus_tc/service_5_event.py
@@ -2,7 +2,7 @@ import enum
 
 from tmtccmd.config.globals import get_global_apid
 from tmtccmd.core.definitions import QueueCommands
-from tmtccmd.pus_tc.base import PusTelecommand, TcQueueT
+from tmtccmd.pus_tc.definitions import PusTelecommand, TcQueueT
 
 
 class Srv5Subservices(enum.IntEnum):
