+++ src/tmtccmd/pus_tm/definitions.py
@@ -1,0 +1,4 @@
+from typing import Deque
+from tmtccmd.ecss.tm import PusTelemetry
+
+PusTmQueue = Deque[PusTelemetry]
