+++ src/tmtccmd/pus_tm/service_3_base.py
@@ -1,4 +1,4 @@
-from tmtccmd.pus_tm.base import PusTelemetry
+from tmtccmd.ecss.tm import PusTelemetry
 
 
 class Service3Base(PusTelemetry):
