+++ src/tmtccmd/pus_tm/service_3_housekeeping.py
@@ -6,7 +6,7 @@
 @author     R. Mueller
 """
 
-from tmtccmd.pus_tm.base import PusTelemetry
+from tmtccmd.ecss.tm import PusTelemetry
 from tmtccmd.pus_tm.service_3_base import Service3Base
 from tmtccmd.utility.tmtcc_logger import get_logger
 from typing import Type
