+++ src/tmtccmd/sendreceive/sequential_sender_receiver.py
@@ -12,7 +12,7 @@ from tmtccmd.sendreceive.tm_listener import TmListener
 from tmtccmd.com_if.com_interface_base import CommunicationInterface
 from tmtccmd.utility.tmtc_printer import TmTcPrinter
 from tmtccmd.utility.tmtcc_logger import get_logger
-from tmtccmd.pus_tc.base import TcQueueT
+from tmtccmd.pus_tc.definitions import TcQueueT
 
 LOGGER = get_logger()
